implement rateLimit using upstash

+++ src/server/api/routers/exercises.ts
@@ -1,5 +1,11 @@
 import { z } from "zod";
-import { createTRPCRouter, publicProcedure } from "~/server/api/trpc";
+import {
+  createTRPCRouter,
+  privateProcedure,
+  publicProcedure,
+  rateLimit,
+} from "~/server/api/trpc";
+import { TRPCError } from "@trpc/server";
 
 export const exercisesRouter = createTRPCRouter({
   getAllById: publicProcedure
@@ -16,59 +22,74 @@ export const exercisesRouter = createTRPCRouter({
         },
       });
     }),
-  create: publicProcedure
+
+  create: privateProcedure
     .input(
       z.object({
         name: z.string(),
         desc: z.string().optional(),
         category: z.string(),
-        authorId: z.string(),
-        authorName: z.string(),
       })
     )
     .mutation(async ({ ctx, input }) => {
+      const { success } = await rateLimit.limit(ctx.currentUser.id);
+      if (!success) throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+
       const user = await ctx.prisma.exercise.create({
         data: {
           name: input.name,
           desc: input.desc ?? "",
           category: input.category,
-          authorId: input.authorId,
-          authorName: input.authorName,
+          authorId: ctx.currentUser.id,
+          authorName: `${ctx.currentUser.firstName} ${ctx.currentUser.lastName}`,
         },
       });
       return user;
     }),
-  update: publicProcedure
+  update: privateProcedure
     .input(
       z.object({
         exerciseId: z.string(),
         name: z.string(),
         desc: z.string().optional(),
         category: z.string(),
-        authorId: z.string(),
-        authorName: z.string(),
       })
     )
     .mutation(async ({ ctx, input }) => {
+      const { success } = await rateLimit.limit(ctx.currentUser.id);
+      if (!success) throw new TRPCError({ code: "TOO_MANY_REQUESTS" });
+
       const updatedUser = await ctx.prisma.exercise.update({
         where: { id: input.exerciseId },
         data: {
           name: input.name,
           desc: input.desc ?? "",
           category: input.category,
-          authorId: input.authorId,
-          authorName: input.authorName,
+          authorId: ctx.currentUser.id,
+          authorName: `${ctx.currentUser.firstName} ${ctx.currentUser.lastName}`,
         },
       });
       return updatedUser;
     }),
-  delete: publicProcedure
+  delete: privateProcedure
     .input(
       z.object({
         exerciseId: z.string(),
       })
     )
     .mutation(async ({ ctx, input }) => {
+      const { success, reset } = await rateLimit.limit(ctx.currentUser.id);
+      if (!success) {
+        const currentTimestampMs = Date.now();
+        const timeLeftInSeconds = Math.ceil(
+          (reset - currentTimestampMs) / 1000
+        );
+
+        throw new TRPCError({
+          code: "TOO_MANY_REQUESTS",
+          message: `Rate limit exceeded. Request rejected. Time Left: ${timeLeftInSeconds} seconds`,
+        });
+      }
       const updatedUser = await ctx.prisma.exercise.delete({
         where: { id: input.exerciseId },
       });
